Hearing Aid
Lead software Engineer/Software Architect
Brief description:I developed a hearing aid for everyone to access by making a custom made audiogram. From the audiogram you are able to find the frequencies that the user can’t hear and store them to the cloud. After you can retrieve the sounds from the phone’s mic and edit it based on the audio profile so the user can hear all the frequencies.